Remove www from git ignore and rename title
This commit removes the www directory from the gitignore file as it is
important to see the source control in this file. Making changes to this
file, whether it be the changing the title as ive done or adding in a javascript
library cnd, is very important in my opinion.