Add book select and cards page
This commit is an important part of the flow of our app. The user will
search for their last read book and they will confirm the book by
clicking on the section. This will push the nav to the cards page and
display the associated books. Added loading svg when searching for
book and no resuls found and also fixed bug where books without a small
thumbnail image would cause an issue in the view.